+++ addons/eu/en16931/bill.go
@@ -1,0 +1,31 @@
+package en16931
+
+import (
+	"github.com/invopop/gobl/bill"
+	"github.com/invopop/gobl/catalogues/untdid"
+	"github.com/invopop/gobl/tax"
+	"github.com/invopop/validation"
+)
+
+func validateBillInvoice(inv *bill.Invoice) error {
+	return validation.ValidateStruct(inv,
+		validation.Field(&inv.Tax,
+			validation.Required,
+			validation.By(validateBillInvoiceTax),
+			validation.Skip,
+		),
+	)
+}
+
+func validateBillInvoiceTax(value any) error {
+	tx, ok := value.(*bill.Tax)
+	if !ok || tx == nil {
+		return nil
+	}
+	return validation.ValidateStruct(tx,
+		validation.Field(&tx.Ext,
+			tax.ExtensionsRequires(untdid.ExtKeyDocumentType),
+			validation.Skip,
+		),
+	)
+}