+++ bill/invoice.go
@@ -266,7 +266,7 @@ func (inv *Invoice) Calculate() error {
 		inv.SetRegime(inv.supplierTaxCountry())
 	}
 
-	inv.Normalize(inv.normalizers())
+	inv.Normalize(tax.ExtractNormalizers(inv))
 
 	if err := inv.calculate(); err != nil {
 		return err
@@ -301,23 +301,12 @@ func (inv *Invoice) Normalize(normalizers tax.Normalizers) {
 	tax.Normalize(normalizers, inv.Payment)
 }
 
-func (inv *Invoice) normalizers() tax.Normalizers {
-	normalizers := make(tax.Normalizers, 0)
-	if r := inv.RegimeDef(); r != nil {
-		normalizers = normalizers.Append(r.Normalizer)
-	}
-	for _, a := range inv.GetAddonDefs() {
-		normalizers = normalizers.Append(a.Normalizer)
-	}
-	return normalizers
-}
-
 func (inv *Invoice) supportedTags() []cbc.Key {
 	var ts *tax.TagSet
 	if r := inv.RegimeDef(); r != nil {
 		ts = ts.Merge(tax.TagSetForSchema(r.Tags, ShortSchemaInvoice))
 	}
-	for _, a := range inv.GetAddonDefs() {
+	for _, a := range inv.AddonDefs() {
 		ts = ts.Merge(tax.TagSetForSchema(a.Tags, ShortSchemaInvoice))
 	}
 	return ts.Keys()
@@ -329,7 +318,7 @@ func (inv *Invoice) ValidationContext(ctx context.Context) context.Context {
 	if r := inv.RegimeDef(); r != nil {
 		ctx = r.WithContext(ctx)
 	}
-	for _, a := range inv.GetAddonDefs() {
+	for _, a := range inv.AddonDefs() {
 		ctx = a.WithContext(ctx)
 	}
 	return ctx