+++ gobl.go
@@ -5,6 +5,7 @@ import (
 	// import all the dependencies to ensure all init() methods are called.
 	_ "github.com/invopop/gobl/addons"
 	_ "github.com/invopop/gobl/bill"
+	_ "github.com/invopop/gobl/catalogues"
 	_ "github.com/invopop/gobl/currency"
 	_ "github.com/invopop/gobl/dsig"
 	_ "github.com/invopop/gobl/i18n"
@@ -19,6 +20,7 @@ import (
 //go:generate go run ./schema/generate.go
 //go:generate go run ./regimes/generate.go
 //go:generate go run ./addons/generate.go
+//go:generate go run ./catalogues/generate.go
 //go:generate go run ./currency/generate.go
 
 func init() {